CA-311: extract UomSelectorSheet into its own file following field+sheet pattern
Moves the bottom sheet out of unit_of_measurement_field.dart into
uom_selector_sheet.dart with a static show() factory, matching the
DateFilterChip/DateRangeBottomSheet convention in the codebase.
Also fixes dynamic l10n param → AppLocalizations and a missing Unit
import in material_cost_form_fields_test.dart from the rebase.
